Added file
The Pacific reef heron (Egretta sacra), also known as the eastern reef
heron or eastern reef egret, is a kind of heron. They are found in many
areas of Asia including the oceanic region of India, Southeast Asia,
East Asia, Polynesia, and in Australia, Tasmania and New Zealand.

Pacific reef herons are medium-sized herons, reaching 57 to 66 cm in
length. They have a wingspan of between 90 and 110 cm and reach an
average weight of 400 grams.

The species displays an unusual, non-sexual dimorphism, with some
members having entirely white plumage and others (the larger portion)
being charcoal-grey. The reason for the colour variation or "morph", is
unknown, though it is most commonly thought to be related to camouflage.

Eastern reef egrets have very short, yellow legs, and the grey variety's
throats and chins are marked by a narrow, white stripe. They have brown
beaks, gold-yellow coloured eyes and the surrounding areas of their
faces are normally of a greenish to yellow cast.

Their food sources are made up predominantly of varieties of ocean-based
fish, crustaceans and molluscs.

The species lay clutches of eggs year round in colonies in the jungle,
between palms and mangroves or in cavities of old buildings. Two to
three paled greenish-blue eggs are laid in nests constructed from
branches and blossoms. Males and females share brooding tasks. They
normally have a 28-day brood period. After chicks are hatched, parents
provide approximately 5 weeks of support.