More Birds, Now Found Randomly
The swallows and martins have an evolutionarily conservative body shape
which is similar across the clade but is unlike that of other
passerines.[2] Swallows have adapted to hunting insects on the wing by
developing a slender, streamlined body and long pointed wings, which
allow great maneuverability and endurance, as well as frequent periods
of gliding. Their body shape allows for very efficient flight, which
costs 50–75% less for swallows than equivalent passerines of the same
size.