Black-bellied whistling duck
The black-bellied whistling duck (Dendrocygna autumnalis), formerly also
called black-bellied tree duck, is a whistling duck that breeds from the
southernmost United States and tropical Central to south-central South
America. In the USA, it can be found year-round in parts of southeast
Texas, and seasonally in southeast Arizona, and Louisiana's Gulf Coast.
It is a rare breeder in such disparate locations as Florida, Arkansas,
Georgia, and South Carolina. There is a large population of several
hundred that winter each year in Audubon Park in uptown New Orleans,
Louisiana. Since it is one of only two whistling-duck species native to
North America, it is occasionally just known as the "whistling duck" in
the southern USA.