Like a Bird on the Wire (Leonard Cohen)
Like a bird on the wire,
Like a drunk in a midnight choir
I have tried in my way to be free.
Like a worm on a hook,
Like a knight from some old fashioned book
I have saved all my ribbons for thee.
If I, if I have been unkind,
I hope that you can just let it go by.
If I, if I have been untrue
I hope you know it was never to you.

Like a baby, stillborn,
Like a beast with his horn
I have torn everyone who reached out for me.
But I swear by this song
And by all that I have done wrong
I will make it all up to thee.
I saw a beggar leaning on his wooden crutch,
He said to me, "You must not ask for so much."
And a pretty woman leaning in her darkened door,
She cried to me, "Hey, why not ask for more?"

Oh like a bird on the wire,
Like a drunk in a midnight choir have tried in my way to be free.