Altamira Oriole
This big tropical oriole is common in northeastern Mexico, but was not
found in our area until 1939. It has since become common year-round in
native woods of far southern Texas. It may go unseen at times as it
forages in dense trees, but it draws attention with its harsh fussing
callnotes. Even before the bird is heard or seen, an observer may notice
its oversized nest, a pouch up to two feet long hanging from the end of
a branch.