Bay-breasted Warbler
This is a characteristic warbler of spruce forest in eastern Canada in
summer. Its numbers vary from year to year, and are likely to increase
quickly during population explosions of the spruce budworm or other
forest pests. This species forages rather slowly compared to most
warblers, moving deliberately among the branches. The male Bay-breasted
Warbler is unmistakable in spring but goes through a striking
transformation in fall, becoming a greenish "confusing fall warbler."