Say's Phoebe
This soft-voiced flycatcher of the west is like the other two phoebes in its tail-wagging habit; but unlike them, it often lives in very dry country, far from water. It is typical of prairies, badlands, and ranch country, often placing its nest under the eaves of a porch or barn. In open terrain where there are few high perches, Say's Phoebe may watch for insects in the grass by hovering low over the fields.