Antillean Nighthawk
A common nesting bird on islands of the Caribbean, this nighthawk enters our area only in southern Florida. When it was first discovered there in 1941 it was considered to be only a subspecies of the Common Nighthawk, as it looks very similar; however, its voice is different. Where Antillean and Common nighthawks meet on the Florida Keys, they appear to compete and to defend territories against each other.