Cape May Warbler
Many of our migratory warblers seem to lead double lives, and the Cape May is a good example. It summers in northern spruce woods, but winters in the Caribbean, where it is often seen in palm trees. In summer it eats insects, but during migration and winter it varies its diet with nectar from flowers and with juice that it obtains by piercing fruit. Birders easily recognize the tiger-striped males in spring, but drab fall birds can be perplexing.