American Redstart
Warblers in general are often called "the butterflies of the bird world," but the Redstart may live up to that nickname more than any other species. This beautiful warbler flits about very actively in the trees, usually holding its wings and tail partly spread, as if to show off their patches of color. At times it feeds more like a flycatcher than a typical warbler, hovering among the foliage and often flying out to grab insects in mid-air.